Extract DNS cache into shared DNS_Cache module
DNS_Cache, DNS_Cache_Record, DNS_Resolve_Data, the resolver worker, and
the validate_ip / get_connected_peer_ip_from_socket / debug_iplisttostring
helpers used to live inside MySQL_Monitor.{hpp,cpp}.  Move them to a new
DNS_Cache.{hpp,cpp} so the same plumbing can back an independent
PgSQL_Monitor cache.

DNS_Cache no longer references GloMyMon directly; counters are wired in
per-instance via set_counters().  monitor_dns_resolver_thread no longer
reads mysql_thread___resolution_family; the desired ai_family is now a
field on DNS_Resolve_Data set by the caller.  MySQL_Monitor wires both up
in its constructor and resolver loop respectively; behavior is identical.

No change to the MySQL DNS cache semantics.
